[change] file structure
